+++ brownie/network/__init__.py
@@ -1,5 +1,6 @@
 #!/usr/bin/python3
 
+
 from .account import Accounts
 from .main import connect, disconnect, gas_limit, gas_price, is_connected, show_active  # NOQA 401
 from .rpc import Rpc